scraper/requirements.txt
requests>=2.28.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
# Optional: for fetch_rendered_html.py (run: playwright install chromium)
playwright>=1.40.0
